resolves #146. only change needed was to rename the file.
