+++ src/cli/info/info.js
